+++ decomp/ST.exe/functions/004DD070/decomp.c
@@ -11,7 +11,7 @@ undefined4 __fastcall FUN_004dd070(AnonShape_004DD070_62E18556 *param_1)
   uint uVar5;
 
   if ((((param_1->field_04D0 == 1) && (1 < DAT_00798f84)) && (param_1->field_04D4 < DAT_00798f84))
-     && ((uint)(param_1->field_04DC + param_1->field_04D8) <= (uint)PTR_00802a38->field_00E4)) {
+     && ((uint)(param_1->field_04DC + param_1->field_04D8) <= PTR_00802a38->field_00E4)) {
     param_1->field_04D8 = PTR_00802a38->field_00E4;
     iVar1 = thunk_FUN_004ac910(&param_1->field_0x1d5,'\x0e');
     /* ST_PSEUDO[raw_pointer_offset]: candidate structure field after proof; otherwise retain buffer arithmetic */
