+++ decomp/ST.exe/functions/004DDD50/decomp.c
@@ -58,7 +58,7 @@ undefined4 __fastcall FUN_004ddd50(int *param_1)
               }
             }
           }
-          if ((uint)(piVar9[2] + local_c) <= (uint)PTR_00802a38->field_00E4) {
+          if ((uint)(piVar9[2] + local_c) <= PTR_00802a38->field_00E4) {
             iVar6 = (&DAT_007e6024)[iVar8 * 5];
             iVar7 = thunk_FUN_004d7b50((char)param_1[9],3);
             if (iVar6 <= iVar7) {
